[FEAT] ParamInfo: Support any submodules
This automatically loads the submodules of fetched param info. This allows
ParamInfo to return all modules available and the submodules for a specific
module instead of just specifically query's and main's modules. With that it is
also possible to get the attributes in general for all modules, which is
necessary to get the prefixes. It deprecates some methods because they were
using the name and not the path.

It also enables users to determine to which module any of the parameters belong
which is currently only possible for submodules of the main and query modules.

Change-Id: I7f5438d8d98649405c9510a610743e4ce4fa7c82