[IMPROV] script_test: Simplify test for stderr
If there is expected output for that script it should just check that. The
additional check if it's an autorun script are unnecessary. The same is true
for the reverse and if there is no expected output and it is not an autorun
script then it should do the checks related to that.

Change-Id: Ic40aee0b1975ad1f785f27bafd3e56bc09511ada