[FIX] l10n_tests: Do lazy network requests
To get the site's language it needs to query the site's siteinfo. When this is
done in a test meta class it will do that regardless before something can check
the net flag of the class. To avoid this it's now requesting the language
inside the test method so that a testing environment can check the net flag
before.

This is also fixing other minor issues in the module:
- Removed senseless comment
- Use 'code' and 'family' attributes in both test classes
- Using super to call the actual super metaclass and not generic type metaclass

Change-Id: Iee5f4ad957b55139a07f711c30e682846df098f3