[FEAT] scripts: Sensible error information
Instead of just printing the help text, scripts now specifically tell
what is wrong and don't print the help (just tell how to do it) so that
the error message doesn't get overlooked in the wall of text.

It additionally returns False if this occurs to make it possible for a caller
of main() to differentiate between success and failure. If it was easy to add
it also returns True if it could be run.

Bug: T91575
Change-Id: I01dc24dc91825446d24a0eb28e563eb9ddb70e87