[FEAT] tools: Add itertools.count function for 2.6
The count function in itertools from Python 2.6 does only support positional
parameters and does not support a step. This uses the count implementation from
the future package just maps it to itertools' implementation in newer versions.

In case an older version of the future package is used (before 0.15.0) it uses
a backported version but informs the user to update.

Change-Id: I779580054642386c90ef2a10a33cfe64cfa1fd23