[IMPROV] tests: add support for hostname check
If no sites are needed but only hostname must be checked it is now possible to
either define one hostname or multiple hostnames and it'll automatically
creates sites from it.

There are already two instances which used a hostname without a site parameter,
which wasn't recognized. And thus they need the net=True explicitly defined.

Change-Id: I4f956c6db6fc4adf62973873759500ec88533423