[IMPROV] replace: Ask for summary if necessary
At the moment it always asks for a default summary unless one was defined via
the command line. But if only fixes are used and all replacements in these
fixes have a summary defined (either via the fix itself or separate for each
replacement) it shouldn't ask for this. This makes it easier to be used
autonomously.

As a byproduct it now knows which replacements from fixes have no summary
defined and lists them before asking for a default summary.

It also adds an option (-automaticsummary) which does prevent that the script
asks the operator what summary it uses and allows to use it without
user intervention but using the automatically generated summary.

Change-Id: Ie72ed81550631c3f53f7991fc688ae73a9b2515e