[FIX] ParamInfo: Correctly initialize ParamInfo
The ParamInfo can do some sanity checking but that is not possible when the
base modules aren't loaded yet. The 424eaa35 change assumed that it would first
add the main modules and then the query modules. But in certain cases it first
added the query modules which let the class think it is properly initialized.
This could only happen when it loads the query modules together with the main
modules and even then could use the working ordering by chance.

Change-Id: I5fb97ba12acf7389c2ce37eff87e1aca432d397a