Lock and unlock Flow topics
This patch allows bots to lock and unlock Flow topics using the API.
It also includes a property for Topic objects showing whether that
topic is locked or not. Post.reply() now raises a LockedPage exception
if the topic is locked. Two tests have been included, one to test
locking and unlocking and the other to test error handling in
Post.reply().

Bug: T108288
Change-Id: I4afb6c0fff6bf541a1a05aa2a5f9b0715e67541e