[IMPROV] i18n_tests: Remove superfluous assertion
If a method raises an exception, its output can't be compared via `assertEqual`
and usually this would lead to the test to error. But if this is then wrapped
in an `assertRaises`, it will expect that exception and thus prevent the error.
But that requires an exception so that the only way `assertEqual` actually does
compare two values will lead to `assertRaises` to fail because no exception was
raised.

This also removes any string formatting which would be done after the method
call when the call is tested and which is thus never done anyway.

Change-Id: I10690b8bd5d263001ae5b12a331b3b465150ca30