[FIX] Request: Ensure that keys are ASCII
When in Python 2 the keys are defined via explicit arguments they are `bytes`
while they can be `unicode` instances when added later or using the
`parameters` parameter or the **-notation.

This is encoding the key using ASCII on Python 2 to normalize the keys and
making the cache easier to compare two requests. If requests with the same
parameters where created using different methods it may not use the same cache
name for all of them.

In Python 3 it's just checking that it's possible encode the keys using ASCII
bot not actually encode it as by default keyword arguments are also named
using `unicode` instances.

The `tests.dry_api_tests.DryCachedRequestTests` are now using the parameters
notation (which was added in b44e59ae) together with having the arguments
explicitly added and using the **-notation.

Change-Id: Idd3ce9588510295ca5742fb31f27ab219f3d2426