[FIX] http: Disable certificate validation
With 952665ac the certificate validation has been enabled for all sites and the
test checking if it can be overwritten wasn't working anymore as the site was
using a valid certificate.

This is reimplementing the feature and ignores the warning emitted that the
validation has been disabled.

This is now working against a site which uses an outdated certificate for
testing purposes. This also removes the test designed to actually highlight
that the certificate is actually valid.

Bug: T105689
Bug: T101138
Change-Id: I0fe9023ca96438c24fdaa71531b17746f09a57e1