[FIX] cache: Wrap items into a list
With 954825e1 the Request class expects that each URL parameter contains a list
to simplify the handling of it. With eedef173 the maintenance script itself
used the new `_params` variable but wasn't actually converting the entries into
lists so that the Request class thought (again due to 954825e1) the strings
itself were lists and thus joined each character.

Change-Id: Ie36a9adcb46675b576226d92e0b2c483b7e2f0d5