[FIX] tests: Don't do hostname check if net=False
When using net=False it shouldn't do any hostname checks as the test is
intended to be run without any network activity. This only affects tests which
also have a site defined (e.g. using dry=True) as CheckHostnameMixin wouldn't
check if no hostname or sites are defined.

Also verifies that net is not False when only hostnames (and not sites) are
defined as hostnames is only there for hostname checks.

Change-Id: Ic60d6d24b489bf63284164411ad5cf1d4ba2ece6