[FEAT] upload: Support ignore warnings callback
This supports a callback method which is executed whenever an upload caused a
warning. This fixes the issue when multiple warnings are reported that the
upload method basically returned only one warning randomly.

It also provides a way to remove the success message as the method now returns
a boolean to indicate whether the upload succeeded.

Change-Id: I391bb2b7fd4a0f318c5b4537d4b615569413f447