Use oauth on beta en.wp and add beta en.ws
Using oauth on beta sites means the 'Pywikibot-test' password
does not need to be used on those sites, which is discouraged as
the beta sites are not secure environments.

Adds oauth to beta en.wp to run user tests on the largest beta site,
providing better coverage on beta sites without a password.

Add beta en.ws to the text matrix, not running user tests,
hoping to catch any bugs on beta when oauth is not enabled.

One test fails on beta en.ws due to the interwiki map (T69931).

Bug: T100797
Change-Id: I607ec16288540d8b6375114ea1ac18e50e6c4828