[IMPROV] site_tests: Separate deprecated calls
The property `nocapitalize` and method `case()` of `APISite` have been
deprecated but this wasn't tested. The test of the `case()` method has been
moved and asserts the deprecation while a new test for `nocapitalize` has been
added.

It also fixed the //instead// string of `case()` which was also adding //use//.

Change-Id: I0285fe24835c2e761092a0aa5379451c06ce10be