[FIX] rollback: Use Revision instance properly
The previous implementation assumed that the revisions are ordered by the id
while they are in fact ordered by the timestamp. It also did iterate over the
revision ids and not over the `Revision` instance which caused the loop to
fail.

Now `APISite.rollbackpage` also requires the user to be logged in.

Change-Id: I280a4781f0f293e09dd94e14007f49cee30f60e3