[FEAT] Fully flexible Site(url)
It completely redesigns how it looks up sites using an URL. Instead of creating
a regex which parses any combination of sites possible it first looks through
the languages in a family to determine which codes use the same domain. After
that it checks if the site uses one the article path or the index.php path.

This uses a Site instance instead of the Family properties to make it fully
flexible and remove the necessity of nicepath. This also removes all regex
related methods in Family.

It fixes a typo in `test_each_family` as it didn't add a separator between the
index.php file and the placeholder. It also removes the test_family from the
skipped entries, as it'll skip test sites anyway (as they aren't in `codes`) so
that it now can match the test family instead.

Bug: T85658
Change-Id: I4cc8bd7b008f0220d87e6a61dd63e4e50194791d