Deprecate fetching all mediawiki_messages using *
Support for fetching multiple messages was added for T60944
in d209142, with b34209d allowing '*' to be used to fetch all messages.

It would be unusual to actually need to use all messages.
More usual would be to need to fetch all message names, which may
be supported using amnocontent.

Move the '*' tests to a dry site, as they are running
particularly slow on ar.wikipedia causing timeouts.

Bug: T111479
Change-Id: Ideefaa198c44bde07d90f23f53991bd3ab81585d