[FIX] Site: Prevent creation of unnecessary Sites
With 0e7ad5ac it needs to do a request to the Site's server to request the
article path. While it'll usually only request it for the Site actually
requested and to any Sites on the same domain, the interwiki map will create a
Site instance for every entry on it. This will then do a request to every
supported site.

This patch is instead just adding a buffer value and instead it only requests
it when the prefix is requested. As it now cannot compare the `Site` instances
in `APISite.interwiki_prefix`, it is now comparing the URLs using the same
technique `Family.from_url' is using.

Bug: T111610
Change-Id: I9d7a7061c9bf621a49867420b2af66b24957ec49