Family defined Site interface
Allow Family class to determine the appropriate Site class
to be used.

It is usable for supporting extensions, using
Site classes composed of several APISite subclasses.

It can also be used for the special case of Sites
which are for a code that is no longer online.

The Family.from_url cache now stores the Family object
instead of the family name, to reduce the number of
Family.load calls.

Change-Id: Ia9f141b9e9f493aa2b1c42afcf4daa8049656c10