[IMPROV] site_tests: Simplify messages test
The test in 21a67b2f was doing three different assertions and the last
assertion required the first two anyway. And the last assertion actually will
produce the most comprehensible output when it fails.

This is now also testing that no deprecation occurs when the special `*` key is
not used.

Change-Id: I1757d3bdad9ac229b838ce6ab35b30bc7520305c