Performance improvement for interwiki.py
interwiki.py makes intensive use of 'pageprops' in order to detect
disambiguation pages. As of now, a separate request is issued for
each page being processed which slows down the processing very
significatly.
This change allows optionally requesting 'pageprops' in site.preloadpages
and modifies interwiki.py to make use of it.

Change-Id: I0b247bfb588caf67cf4799b6ee566e39552717ac