Remove wikidata cosmetic changes flag in page.py
73273de4 introduced a workaround in page.py to prevent
cosmetic changes being activated on Wikidata.  At the time,
the 'wikidata' family had two sites, 'client' and 'repo'.

Since early 2013 with 03e87f2d the code for the production
repository changed to 'wikidata', when Wikidata moved to wikidata.org
However the code used in the workaround in page.py was not updated,
so this workaround has not been effective for over 2.5 years.

Change-Id: I85cdfbf39c310279fabd2aea1ddbbb3e3d47e14b