Reduce import dependencies around cosmetic_changes
There are several imports around cosmetic_changes which depended
on each other, causing complications when adding new imports.

pywikibot.cosmetic_changes imported pagegenerators only for
parameterHelp, which is unused since the module was moved into
the pywikibot package in fd1ba1e48e.

As cosmetic_changes imported scripts.isbn for backwards compatibility,
it inherited isbn's dependency on pagegenerators.  The import of
scripts.isbn has been moved into the function where it is used
to delay the import until after initialisation.

As pagegenerators depends on wikidataquery, which depends on page,
the wikidataquery dependency has been moved to be a delayed import
within WikidataQueryPageGenerator.

Also move the import of page module in cosmetic_changes down to
where it is used, so that cosmetic_changes doesnt depend on page
during startup.

Change-Id: I5ec082c65247129454b605781210f09c168b4a51