[IMPROV] NamespacesDict: Lookup names from dict
Instead of iterating over all namespaces, the NamespacesDict can cache a
dictionary of names mapping to the namespace instance.

This cache is now used by `validLanguageLinks` which also avoids to iterate
over all namespaces for each langlink and to avoid any magic provided by the
`Namespace` class.

Also add a possibility to lookup the namespace via an already normalized name.

Change-Id: Ice7f51aef1673724c87ae2aa7a86f35f36d01692