[FEAT] upload: Get stash info before continuing
If the response does contain warnings it won't contain the offset for the
continuation (T73737). So it requests the stash info for the given file key to
determine the offset. This also allows to properly reuse the file key as it
will create a new file key instance when the offset is 0. It also prevents
that the API shows an error if the offset is neither 0 nor the actual size
uploaded.

And if the offset is given it can verify that it matches the stashed data. It
also supports comparing the SHA1 of the already uploaded part to minimize the
chance that a different file is uploaded for continuation. It now also verifies
that the offset is not larger than the file size. In case the file is large and
doing an SHA1 would be time consuming it's possible to not verify them.

The offset is not present if the warning occurred in the first chunk which
would have caused that the first chunk is discarded when continuing to upload.
To do that the file key must be defined before it issues the warnings.

Also when uploading in the unchunked mode and a file key is provided it did
upload the complete file even though the offset was set to `False` to indicate
the file is completely uploaded.

One test is an expected failure due to T112416.

Change-Id: I610a0744d74dc5a2523d96eabde818f76dd6c5e8