[IMPROV] site_detect: Fail with more specific error
When the API from a private wiki isn't exposed it may be able to determine the
API path but it may not be able to determine the articlepath. This is guessing
the article path from the original URL (if it wasn't the API URL) and also sets
the language to the family name as a fallback.

Bug: T112575
Change-Id: Ib7bc898ea179c5806a40260dbb69abf2084ed973