[FEAT] Support writing to any compressed file
Instead of just opening the file to read it, this also allows to write them. To
have a similar naming scheme like `open` it renames the function to allow
changing the signature.

It now by default uses the file extension and not the magic number when using
the new function. And the category script now uses the new function.

Bug: T112373
Change-Id: I24f5477c2488387cd7811e4c38859323720aef88