[FIX] tests: Correctly handle abstract classes
The condition for an abstract test class was just that there are no test
classes in the current class it tries to create. So when a test case
subclassed another test case and didn't add more tests it was classified an
abstract class and it skipped a lot of checks, which includes whether it's a
networked test.

Also fix the `MementoTestBase` class as it subclasses just `TestCaseBase` which
doesn't do any magic like automatically determining `net` when for example a
`hostname` is set or subclassing `CheckHostnameMixin` when a `hostname`
attribute is set.

Change-Id: I0bdf4d1ee63e5123b519ba6631a79514bf3d8a7c