[FIX/IMPROV] Add option class to output the list
The alternatives output by `listAlternatives` are zero-based while the option
is one-based since 9c590fed. This adds the option class `OutputProxyOption`
which just calls `output` of the class used for it.

One class which can be used is `SequenceOutputter` which can be used together
with `ListOption` and does just create a string listing the contents of a list.
That class is creating a one-based string which is output. It also has a
flexible indentation depending on the number of alternatives.

Using the separate `SequenceOutputter` class has the advantage that it's
possible to create such strings without being tied into using an option.

Bug: T107272
Change-Id: If843cb0af86292b6a6ab426662d4831503ea47b5