[FEAT] color_format to ignore valid colors
This allows to add colors to strings which will be filled using format. Instead
of calling 'str'.format(…), color_format('str', …) is called. That uses the
`_ColorFormatter` class which parses the text differently and skips color
fields.

It also checks that the color fields are used without conversion and format
spec. The `\03` marker becomes optional and is added if missing and if it may
not be present anywhere else than in front of color field.

Change-Id: Ib354fb80180b0e421a18474cc67d7ee10cda3745