tests: replace sys with tools
Use existing tools PY2 and PYTHON_VERSION.

While performance is not a consideration, these variables are
shorter and more descriptive.

Use new tools StringTypes and UnicodeType, are equivalent to the
types.StringTypes and types.UnicodeType that were dropped in Python 3.
These reduce the need for importing sys and doing version specific
logic at the beginning of each module.

Change-Id: Ie72f8df1dabe988d62bbc6337512df6b5daf47b7