ui_tests: limit retries in testOutputUnicodeText_no_transliterate
The old code could loop infinitly. The maximum number of retries
is now limited to 3 (= at least 3 seconds) which should be enough
for all practical purposes.

The test was adapted from requiring 'Alpha' on the first line to
requiring 'Alpha' somewhere in the return value, as output before
the first expected line is a regular issue.

In addition, changed incorrect ctrl-L to alt-L to select the font
selector. This still probably only works for a Dutch Windows
version, so we might want to consider doing this either manually
or with a direct Win32 API function (assuming we won't test on
WinXP)

Change-Id: I87e40f0c200a5cbd181df5e4112055e38650414a