[FIX] Implement getargspec in tools
With version 3 inspect.getargspec has been deprecated and with version 3.6 it
is supposed to be removed. This implements getargspec in tools which just calls
the implementation before Python 3.3. In Python 3.3 it calls the replacement,
inspect.signature, and constructs a result simulating getargspec from it.

It removes the special case for handling deprecation warnings in Python 3.5
introduced in 773e72ee.

Bug: T106209
Change-Id: Ib68f09b3e47169279d9725ae0ddfab71009b2e1c