[FIX] script_tests: Remove weblinkchecker dependency
When the `memento_client` dependency for `weblinkchecker` was introduced in
3bf9bfac it was added to the script tests. But the script runs fine without it
when it's not installed and causes an unexpected success. But this didn't cause
a problem because the script included the file name extension so it wasn't
expecting a failure as it was expecting it from a script with an additional
`.py` extension.

Change-Id: I8510cfa2553d5d5ad7e04c633b60b0a7ed9687be