[FIX] patrol: Replace mwlib with mwparserfromhell
This removes the dependency of `mwlib` (which doesn't work on Python 3)
and uses `mwparserfromhell` instead. Unfortunately does the `mwparserfromhell`
implementation not support lists directly so the parsed result has to be
filtered manually.

It now allows links outside of lists and normalizes the titles and checks
against the aliases for `Special:PrefixIndex`.

Bug: T95142
Change-Id: I4f7b5c7a67e0c90530319fce1f3ab0ca0c1a1138