Remove dependency on doctest-ignore-unicode
doctest-ignore-unicode is used to ignore u'..' prefixes
in the doctest output.

When it isnt installed, doctest complains with an uninformative
error:

ValueError: line 34 of the doctest for api.Request has an invalid
option: u'+IGNORE_UNICODE'

Instead of using doctest-ignore-unicode, downgrade the values to
a `str` to remove the u'..' prefix on Python 2.

Change-Id: I4471a3eb658a98970b49dd5e3d71e1d5e195d0e6