[IMPROV] Provide helper for generated tests
With 4ec75cd7 all tests which were generated using multiple sites are using
`__name__` now. But there are other cases which dynamically create tests and
may not add that property. So `MetaTestCaseClass` now supports `add_method`
which adds the value to the `dct` and sets `__name__` and allows to appends
something to a docstring (or to overwrite the docstring).

Change-Id: I7797ee461d82d764c32e92f773e9c7433fda24bd