[FEAT] plural_tests: Verify plural rules
This verifies the consistency of all plural rules. Static plural rules must
have `nplurals` set to 1 and return 0 for `plural`. If the plural is a callable
it'll test all values from 0 until `max_num` and check that the indexes
returned are greater than 0 but also below the number of plurals. That maximum
number thus must then cover all possible plurals for all languages.

Change-Id: If2a665fa047630de30ac51a108144affb08afa16