[IMPROV] i18n: Replace all plural items at once
Instead of generating one list of plural items and to substitute always the
first occurrence it can actually iterate over the string using `re.sub` and
dynamically replace it with a plural replacement.

It does verify that `nplurals` is 1 when the `plurals` is not callable. And if
`nplurals` is 1 it verifies that that the value returned is 0.

Change-Id: Id72f7bd55ddd816f6d15ba69e9b8e936333625fd