[FIX] cosmetic_changes: Implement external link to wikilink
The feature to replace external links into a wikilink when it pointed to the
same wiki was introduced in compat as of f2645f85. But it had been disabled
since its addition. Later it was, still disabled, ported to core in 6ac688fe.

This now enables an improved version which does not allow the delimiter in the
text itself so that links to diffs for example won't be changed. Also instead
of using the assumption that the URL is `code.family` it's iterating over all
the URLs which are also available for `Family.from_url` and `Site(url=…)`.

Change-Id: Id06b256e2005f5730fbfbaa96cd6690ec4de1789