[FIX] config: Don't crash on later get_base_dir calls
All module variables starting with only one underscore get deleted after the
module's instantiation and with 2b07db2a the variable to store whether no user
config should be loaded was stored in such a variable and thus deleted so that
when `get_base_dir` was called afterwards it would crash because it is missing.

This patch renames the variable into using two underscores to still hide it's
presence but prevent it from deletion after the instantiation.

Change-Id: I538a73ff37da27a34ab732becafbf5a58b72a464