[FIX] Prevent string comparison for Coordinate
When it tries to compare a string to `Coordinate` it previously parsed the
string and returned True if the string matched the coordinate. Otherwise it
just continued to work on it and compared the object with the string directly
which never evaluated to True, so it always returned False.

But with 4f4770f3 `Coordinate` implements `__eq__` and now that method tries to
compare a string to a `Coordinate` and that raises a `AttributeError`. So
instead of just returning True and falling back to the default comparison
otherwise it directly returns False if the comparison with the parsed string
failed.

Change-Id: Ibe322383a319c57fd20db43758f87864550ea0e4