[IMPROV] Yield interwiki URLs separately
Instead of returning one list with all URLs it can yield one URL at a time.
This way it's possible to only query the article path if nothing else matched
first.

This also adds a parameter to avoid manually skipping the first entry which
isn't actually a suffix for an article URL.

Change-Id: Idd20f0757856299d2ae21608bb29d12c3e4cae11