[FEAT] cache: Separate commands
Instead of having everything in one command and to play with `and` to only
execute part of the command, this is providing three separate commands to
filter for entries, output something depending on them and to do an action
afterwards.

The filter command only needs to return a boolean for an entry. The output
command can either return None to suppress the default output, similar to
before. If it returns anything else it'll print that result. And at the moment
only the delete action is supported and that as an option and not using code
like the other two.

Additionally it supports to only define a function name and it'll call it
automatically with the entry as the first positional parameter.

It adds two utility output functions to easier see what each cache entry is
actually doing.

Change-Id: I7b8fc283555a6d35719e47d28ff1becfbd07baf7