[FIX] Proper deprecation for old DataSite.get_*
This removes the explicit deprecation of `_get_propertyitem` and adds a
non-deprecated `_get_item` which can be called from `_get_propertyitem`.

So whenever an old `get_*` method (like `get_info`) is called it actually
returns a deprecation for that and not for `_get_propertyitem` and `get_item`.

It also avoids using lambda by replacing it with functools.

Change-Id: I71e8eb4e7df3f2ec7ad9aad39d10be88ace4ed42