[FEAT] site: Dynamically determine Wikibase repo
Instead of statically defining the Wikibase repository in the family, it can
query the API to determine the repository without user input. This also changes
the implementation in `Family` to use the site's implementation.

This does not provide the `DataSite` instance if there is no such family which
supports that URL. And while previous versions would raise a
`SiteDefitionError` if the data returned by the family is invalid, this
implementation does catch the error as it assumes the information is valid and
it instead means that there is no such family created yet.

Bug: T85331
Change-Id: I5568d848a369c8af36a40c193e2409071e23a5fb