[IMPROV] TimeStripper: Deprecate findmarker duplicate
The `findmarker` method in `TimeStripper` works exactly the same as the
function in the module. This is deprecating the method in the class in favour
of the module function.

This is also introducing a new super class in the corresponding tests which
will take care of creating the `TimeStripper` instance.

Change-Id: Ia349e78e68d94fc6c6ceab75d8b679bbee48b2b7