Use new test topic for reply tests
The topic on the testwiki used to test replying has become too big.
The size of the topic is slowing down reply tests. This patch points
the reply tests to a new topic in order to speed them up.

Unless someone can purge the posts from the topic, patches such as
this one will be necessary periodically.

Change-Id: Icf84daa1ec0eb855370e10902b82cbb43ae7af43