Set url to None if it contains only "*" in the weblink argument
When "-weblink:*" is the argument passed, "*" is passed as the euquery url
to the MediaWiki API which raises a geubad_query: Invalid query. So, a
check is made on url in handleArg() if the argument starts with "-weblink"
and if it is a * then it is set to None. Now, the MediaWiki API will
generate every page with any URL.

Bug: T113142
Change-Id: I3a324942f6ac4a187179a6db5584361d65d71e04