[IMPROV] Page repr tests: Enforce console encoding
When backporting 38589d3 to the 2.0 branch, an error occurred on the Jenkins
tests because the console encoding wasn't the expected value. Now for some
reason this is not a problem in the master branch, but in case someone uses a
custom console encoding it might fail even though it shouldn't. So this is now
enforcing the UTF-8 console encoding.

Change-Id: Ia642c34b67bf7edd0da9240a80552911b9e79918