solve_disambiguation.py: Correct the way main namespace is detected
Previously, passing "-main" option caused the "ignore_title" dict to extend
and include namespace numbers. But those numbers never match a page.title().
Hence the non-main-namespace pages could not be detected.

In the new implementation the "ignore_title" is not extended, instead
"namespaces=0" is passed as a keyword argument to getReferences.

Bug:T105892
Change-Id: Icde898746dd77309d3e42b2da6c40f9bc4ab1194