Fix edit conflicts after calls to save_claim()
Unlike DataSite.addClaim(), DataSite.save_claim() was not updating
the latest revision ID of the item of the saved claim. As a result,
any submission to Wikidata after modifying a claim, such as
Claim.changeRank(), would result in an edit conflict.

Now DataSite.save_claim() updates the latest revision ID of the
item of the saved claim.

Change-Id: I2079228b50abca2e171bbc63e6bdd581532cd703