imageuncat: Deprecate -yesterday for -logevents
The cli argument -logevents uses LogeventsPageGenerator to find
pages based on logevent. This can be used for imageuncat instead
of the -yesterday cli arg. But it doesn't filter based on
timestamp.

This commit adds the start and end timestamps to -logevents and
the same args to LogeventsPageGenerator. Hence, this now deprecates
the older -yesterday cli arg in imageuncat.

Bug: T67192
Change-Id: I5cd14609a43bd7b7343f234c0832cc471ae0250b