[bugfix] Enable script_tests for featured.py
featured -simulate doesn't response "'Use -help for further information."
but there are 4 lines:

  Link FA does not exist
  <empty line>
  NOTE: featured articles are not implemented at wikipedia.de
  0 pages written.

We cannot use first or 3rd line because of the localized answer for
template and site. We can expect the last line as response.

The behavior of the response might be changed because the data is hold
by wikibase for wm sites.

Bug: T128955
Change-Id: Ia9ac862d2d07db47b1edc87c89aa4136c524891c