site: Support async chunked uploads (T129216)
Async chunked uploads (MW >= 1.21) will tell the server to defer jobs
such as assemble chunks and publishing files to the job queue, so that
it would not suffer from the very short timeout of HTTP requests. This
will enable larger files (more than one gigabyte) to be uploaded by
pywikibot without failing after 503 (T128358) like non-async chunked
uploads.

The mechanics is that when a such job is submitted, in queue, or in
progress, the MediaWiki server shall send a result='Poll' response.
Polling means to periodically send an action='upload', checkstatus=True
request and handle the result accordingly, with the same logic as
in the response that initiated the poll.

In addition, handling result in responses has been seperated into
if-elif cases to avoid irrelevant warning messages in logs. (T133443)

The upload.py script has learned a new flag (-async) which is passed
to the APISite upload function.

Bug: T129216
Bug: T133443
Change-Id: If9e3330c2f4065e491f20475bee0c7da7ccbd096