pywikibot.tools: Raise error if bz2 not found
bz2 is required in pywikibot to open compressed files. But,
sometimes python is not compiled with bz2. This causes pywikibot
to fail while it is being imported.

This commit shows a warning in such cases and tries to use
bz2file as an alternative. If both are not found, it raises an
error lazily (i.e. only when bz2 is needed to open a compressed
file).

Bug: T123092
Change-Id: Iace28a5b356cff19118983fd089770c75a1b4dff