Use flake8-putty 0.3.2 auto ignore
Instead of flake8's general # noqa, flake8-putty 0.3.2 allows
individual codes to be disabled with '# flake8: disable=xxx'

The noqa in generate_*.py regarding PYWIKIBOT2_NO_USER_CONFIG
were to prevent E402 (module level import not at top of file)
however this is already unconditionally ignored by flake8 and pep8.

The noqa in editor.py was added in 728524b, when it was an
unused variable.  It is now used, so no error is generated.

Change-Id: I21ee56755311c4c5f9cd427e2ce8aca35568b7f7