[bugfix] Fix deprecation warning for namespace shortcuts
namespaces.CONSTANT gives the Namespace object not the string.
Assigning Namespace to a string adds a colon and prepends one for
FILE and CATEGORY Namespace because Namespace.__str__ uses the
canonical_prefix method whereas the old shortcuts gives the default
localized name which is custom_name property. The deprecation warning
should follow this issue.

Change-Id: Idaa5e28dccbdd0311b617ec8607c4bbc96d88fa7