Add wikisourcetext.py: load text in Page ns
This bot applies to wikisource sites to upload text.

Text is uploaded to not-(yet)-existing pages in Page ns, for a specified
Index.

Text to be stored, if the page is not-existing, is preloaded from the
file used to create the Index page, making the upload feature independent
from the format of the file, as long as it is supported by the
MW ProofreadPage extension.

Change-Id: I42eff9ae6f0cc4170381dcb1ef04869252662493