Activate nosetrim to reduce repeated errors
The OAuth jobs for wmfbeta sites are failing due to an authentication
problem, causing mosts tests to fail.

Nose normally and helpfully emits a lot of logging information for failed jobs.
However when there is a systemic problem, as currently occurs with the
wmfbeta sites, the result is the job logs are very large, which affects
developers ability to understand the failures and it exceeds the Travis
maximum log size.

Change-Id: I048e04a9451d25bf6893755bf73007f944578191