Replace ffind with findx
ffind (introduced in 7c6c3fd4c8) is failing on Jenkins since June 2016,
for an unidentified reason.

As this is preventing all development, a workaround is needed.

findx provides similar functionality, except that it depends on
Unix-style binaries find and xargs, which is a regression, and
doesnt work on Python 2.6.  However these problems are not
relevant to the Jenkins environment which is Unix and doesnt
support Python 2.6.

Bug: T137628
Change-Id: Iedfc0bf3a1a99d904aa3464a30eb6b2f594ed28d