WikidataSPARQLPageGenerator should not always remove duplicates
get_items creates a set from the results, to avoid duplicates. It has
the unfortunate side effect of shuffling the results of all queries,
including those with an ORDER BY.

The result_type argument is added to allow the caller to change the
default set() to a list() to preserve order and duplicates.

Bug: T141348
Change-Id: I19873439e82f1a74797d1d2b2ac253086b38721e