Require site object only for older MediaWiki installations
The patch in Ifbc3fc0f324c7483f0eba092529a05a8f617ec8c introduced
the need for providing a Site object with WbQuantity to ensure
correct handling of missing error bounds. This affects anyone
working with Wikidata.

This patch changes the behaviour so that Wikidata users can continue
as before and Site objects are only needed for users of older
MediaWiki installations. It also ensures that Wikidata users do not
accidentally go on assuming +/-0 error bounds as default.

See T150210 for discussions.

Bug: T150210
Change-Id: I406db1e86dff772dae7af557859f1458ca87ba5d