FilePage.get_file_history(): Check for len(self._file_revisions)
Previously it was hasattr(self, '_file_revisions'), which will always
return True due to attribute being set on instance initialization.

Unlike other cacheing logic such as templates or categories that use
hasattr check, we keep this logic of len check due to the fact that
a file page must have a file revision, and many methods expects the
self._file_revisions dict to be non-empty.

Bug: T155740
Change-Id: I4f35971a3f45f6ffed20135f45c5050f862a0d92