site: Handle offset mismatches during chunked upload
If the server returns an error code of 'stashfailed', error JSON
contains an offset, and the provided offset is different from
the offset we have sent it, we believe the server refuse to accept
this chunk because of offset mismatch.

A condition is added to APIError handling so that if such error is
encountered, we update our own offset and continue the upload.

Bug: T156402
Change-Id: I78fff13769e887ce554e8acf8664b1ce8e8bb5f5