APISite.data_repository(): handle warning with re.match()
In 4e6810e, the message was added to i18n. During the process,
the message changed: a period was added to the end, and the single
quotation marks (') were changed to <var> tag, which in turn creates
double quotation marks ("). However, luckily, the warning (somehow
converted from an error internally in MediaWiki) seems to stay in
English regardless of which site we are operating on.

Like the other places where req._warning_handler is used, here, to
maintain backwards compatibility, we regex-compare the warning to
account for the changes to the warning message. The regex itself is
not wrapped after 80 characters due to that wrapping a string would
be easily confused with wrapping multiple parameters.

Bug: T156596
Change-Id: I47436809960de23746eeb4877a9919772ce11448